[codec-memcache] Implement ASCII Memcache Codec.
This changeset adds support for the ASCII (aka "old") memcache protocol. Because of the diverse
nature of the different commands, it is not as elegant code-wise than the binary codec. All commands
as defined in the spec are implemented (aside form SLAB assignment/stats).

Note that the codec just implements the encoding and decoding of messages, is does not do any kind
of sanity checks if the message sent makes sense semantically (since custom implementation can
extend the protocol at will). Proper memcache (or memcache-compatible) APIs can be built on top
of it.

todo:
	- test store request with payload
	- test retrieve response with payload (end!)

	- add request decoder (+ tests)
	- add response decoder (+ tests)
		.. abstract decoder should handle the line termination thing and abstract stuff, the actual impl
			the meaning of the lines ...
	- add ascii aggregator

	- refactor responses to make the enums static .. for the easy ones.
	- version.. also make it static so no new object all the time
	- refactor noreply (and maybe others into toplevel obj)
	- refactor the encoders afterwards to be more concise and reduce repetition
	- add proper documentation, file headers,...